Fix parallel runs with reportable caching crashing
Reportables have a record attr, which contains a reference to manager,
so pickling an entire reportable tries to also pickle the manager and
the multiprocessing Lock instance if it exists. Pickling this lock
causes crashes. This was fixed by the caching process creating a shallow
copy of the reportable to be saved and setting the record to `None`,
since it will be repopulated on reportable reload anyway.