Add RANK env var checks for parallel mode
Running pytorch with distributed strategy can cause issues if it re-runs
the main script multiple times. Curifactory creates a report and
experiment run for each individual process, and a race condition on
files like the manager store can occur. We added in a LOCAL_RANK and
NODE_RANK check in the experiment run, which will automatically set
parallel-mode (don't touch store or create reports) for all except RANK
0 processes.

Additionally, we fixed a bug where parallel-mode would only work if a
global index was specified, which only applies when curifactory is run
with the `--parallel` flag, rather than some other form of parallel
processing implementation as in the above case.