Add path tracking code in record and staging
Record now contains a list of paths retrieved through get_path and
get_dir, which the staging store_outputs now uses to copy any manually
used paths to store full folder. We chose doing this rather than keeping
a list of paths on the manager because a manager doesn't necessarily
have an "end" state, and stages are already themselves in charge of
managing re-saving objects to the full store.