Add cacher argument dictionaries for pandas calls
If any necessary escapechar or other attributes need to be set on the
PandasCSVCacher and PandasJSONCacher, they can be passed in as argument
dictionaries. This commit potentially fixes #4. Additionally, this
commit fixes a problem where the PandasCSVCacher by default doesn't
correctly reload a dataframe because it fails to account for the index
column.