Fix full store distributed run doing full store on all processes
Only non-rank-zero manager has store_entire_run set. A warning for the
user was added so they know this is the expected functionality.
Technically it's not an _error_ to store full for all of them, but it's
a ton of data redundancy, and this is problematic in runs with lots of
data.