Add normal cache check after DAG-execution check in stages
There are cases where the same stage might get called multiple times
with the same arguments - the DAG will only see that it doesn't have
the cached values at first, and since it doesn't update as the
experiment is running, it will think every subsequent call of that stage
will still need to execute. By adding in the cache check even when
the DAG things the stage should execute, we ensure that it can still use
cached values from earlier in the experiment and skip execution.